+++ CHANGELOG.md
@@ -11,6 +11,12 @@
     - Use `text` instead of `reply`
   - Add support for `generation_id`
 
+## 4.1.8
+
+- [#206](https://github.com/cohere-ai/cohere-python/pull/206)
+    - Update cluster endpoint to use UMAP+HDBSCAN
+    - Remove threshold and add n_neighbors and is_deterministic as params
+
 ## 4.1.6
 
 - [#205](https://github.com/cohere-ai/cohere-python/pull/205)